nlp-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
pydantic==2.4.2
